% This file is maintained manually for mand 3.0.
% It preserves the published mand 2.0 data documentation.
% Please edit documentation in R/src.r
\docType{data}
\name{atlas}
\alias{atlas}
\title{Atlas set}
\format{
A list of array
}
\usage{
data(atlas)
}
\description{
The data is the atlas image data. An image whose element is "ROIid" is stored for each atlas.
}
\keyword{datasets}
