% This file is maintained manually for mand 3.0.
% It preserves the published mand 2.0 data documentation.
% Please edit documentation in R/src.r
\docType{data}
\name{atlasdatasets}
\alias{atlasdatasets}
\title{Atlas data set}
\format{
A list of matrix
}
\usage{
data(atlasdatasets)
}
\description{
The data is the atlas data. Various atlases are stored. Each matrix has "ROIid" and "ROIname" as column names.
}
\keyword{datasets}
