% This file is maintained manually for mand 3.0.
% It preserves the published mand 2.0 data documentation.
% Please edit documentation in R/src.r
\docType{data}
\name{baseimg}
\alias{baseimg}
\title{Base Brain Data}
\format{
A array
}
\usage{
data(baseimg)
}
\description{
The data is the base brain data. This is an average image of a healthy person, and is used when generating artificial data.
}
\keyword{datasets}
