% This file is maintained manually for mand 3.0.
% It preserves the published mand 2.0 data documentation.
% Please edit documentation in R/src.r
\docType{data}
\name{exbrain}
\alias{exbrain}
\title{Example Brain Data}
\format{
A array
}
\usage{
data(exbrain)
}
\description{
The data are from a MRI gray matter brain data for one subject.
}
\keyword{datasets}
