% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-api-legacy.R
\docType{methods}
\name{imgdatamat}
\alias{imgdatamat}
\title{Creat Data Matrix Function}
\usage{
imgdatamat(
  imgfnames,
  mask = NULL,
  ROI = FALSE,
  atlas = NULL,
  atlasdataset = NULL,
  ROIids = NULL,
  zeromask = FALSE,
  schange = FALSE,
  ...
)
}
\arguments{
\item{imgfnames}{a vector for (nifti) file names to be used.}

\item{mask}{a vector for brain mask data.}

\item{ROI}{a logical for roi data set.}

\item{atlas}{An optional three-dimensional atlas image.}

\item{atlasdataset}{a matrix or data.frame. colnames shold include "ROIid" and "ROIname".}

\item{ROIids}{a vector}

\item{zeromask}{a logical for masking voxel with all zeros.}

\item{schange}{a logical for change dimension.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\item{S}{data matrix}

\item{brainpos}{binary brain position.}

\item{imagedim}{three dimensional vector for image dimension}
}
\description{
This is a function that creates a data matrix for analysis from a file saved in image format.
}
\details{
\code{imgdatamat} requires image file names.
}
\examples{
# imgfnames1 = c("img1.nii", "img2.nii")
# imgdata = imgdatamat(imgfnames1)

}
