% This file is maintained manually for mand 3.0.
% It preserves the published mand 2.0 data documentation.
% Please edit documentation in R/src.r
\docType{package}
\name{mand-package}
\alias{mand-package}
\title{Multivariate Analysis for Neuroimaging Data Package}
\description{
A Package for implementation of multivariate data analysis for neuroimaging data.
}
\references{
Kawaguchi, A. (2021). Multivariate Analysis for Neuroimaging Data. CRC Press.
}
\author{
Atsushi Kawaguchi. \email{kawa_a24@yahoo.co.jp}
}
\keyword{documentation}
