% This file is maintained manually for mand 3.0.
% It preserves the published mand 2.0 data documentation.
% Please edit documentation in R/src.r
\docType{data}
\name{mask}
\alias{mask}
\title{Brain Mask}
\format{
A array
}
\usage{
data(mask)
}
\description{
The data is the brain mask. This is used to exclude extra-brain regions from the analysis.
}
\keyword{datasets}
