% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-api-legacy.R
\docType{methods}
\name{multicoat}
\alias{multicoat}
\title{Multi Coat Function}
\usage{
multicoat(imgs, y = NULL, row4imp = 6, col4imp = 1, trm = NULL, ...)
}
\arguments{
\item{imgs}{A list of three-dimensional image arrays to be displayed.}

\item{y}{image(s) to be overcoated. In the case of a single image, the same y images are overlaid on each x. In the case of a list of images equal to the number of x, each image is overlaid on each x.}

\item{row4imp}{the number of rows per a image}

\item{col4imp}{the number of columns per a image}

\item{trm}{the index to trim the top and bottom of the slice}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This is a function for plotting an image. The analysis result can be overcoated on the template.
}
\details{
\code{multicoat} requires a image array.
}
\examples{

data(exbrain)

}
