% This file is maintained manually for mand 3.0.
% It preserves the published mand 2.0 data documentation.
% Please edit documentation in R/src.r
\docType{data}
\name{sdevimg}
\alias{sdevimg}
\title{Standard Deviation Brain Data}
\format{
A array
}
\usage{
data(sdevimg)
}
\description{
The data is the standard deviation brain data. This represents the common standard deviation between the average images of healthy subjects and patients with Alzheimer's disease, and is used when generating artificial data.
}
\keyword{datasets}
