% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-api-legacy.R
\docType{methods}
\name{sizechange}
\alias{sizechange}
\title{Size change Function}
\usage{
sizechange(img1, simscale = NULL, refsize = NULL, ...)
}
\arguments{
\item{img1}{a array or nifti class, which is a image data to be changed the size.}

\item{simscale}{a numeric.}

\item{refsize}{a vector with length 3, which is a size to be changed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
TThis is a function that changes the resolution of the image.
}
\details{
\code{sizechange} requires the array data.
}
\examples{

data(exbrain)
exbrain2 = sizechange(exbrain, simscale=1/2)

}
