% This file is maintained manually for mand 3.0.
% It preserves the published mand 2.0 data documentation.
% Please edit documentation in R/src.r
\docType{data}
\name{template}
\alias{template}
\title{Brain Template}
\format{
A array
}
\usage{
data(template)
}
\description{
The data is the brain tempalte. This is an average brain image, and is mainly used for overlaying analysis results.
}
\keyword{datasets}
