% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{data}
\name{atlas}
\alias{atlas}
\title{Atlas set}
\format{
A list of array
}
\usage{
data(atlas)
}
\description{
The data is the atlas image data. An image whose element is "ROIid" is stored for each atlas.
}
\keyword{datasets}
