% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{data}
\name{atlasdatasets}
\alias{atlasdatasets}
\title{Atlas data set}
\format{
A list of matrix
}
\usage{
data(atlasdatasets)
}
\description{
The data is the atlas data. Various atlases are stored. Each matrix has "ROIid" and "ROIname" as column names.
}
\keyword{datasets}
