% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{data}
\name{diffimg}
\alias{diffimg}
\title{Difference Brain Data}
\format{
A array
}
\usage{
data(diffimg)
}
\description{
The data is the difference brain data. This represents the difference between the average images of healthy subjects and patients with Alzheimer's disease, and is used when generating artificial data.
}
\keyword{datasets}
