% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{package}
\name{mand-package}
\alias{mand-package}
\title{Multivariate Analysis for Neuroimaging Data Package}
\description{
A Package for implementation of multivariate data analysis for neuroimaging data.
}
\references{
Kawaguchi A, Yamashita F (2017). Supervised Multiblock Sparse Multivariable Analysis with Application to Multimodal Brain Imaging Genetics. Biostatistics, 18(4) 651-665.
}
\author{
Atsushi Kawaguchi. \email{kawa_a24@yahoo.co.jp}
}
\keyword{documentation}
