% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{data}
\name{mask}
\alias{mask}
\title{Brain Mask}
\format{
A array
}
\usage{
data(mask)
}
\description{
The data is the brain mask. This is used to exclude extra-brain regions from the analysis.
}
\keyword{datasets}
