% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{methods}
\name{rec}
\alias{rec}
\title{Reconstruction}
\usage{
rec(Q, imagedim, B = NULL, mask = NULL)
}
\arguments{
\item{Q}{a vector for reduced data.}

\item{imagedim}{a vector for original dimension.}

\item{B}{a list or a matrix indicating the basis function used in the dimension reduction.}

\item{mask}{a list or a matrix indicating the mask image used in the dimension reduction.}
}
\value{
\item{}{a reconstructed array}
}
\description{
This is a function that restores the vectorized image to its original dimensions, reduced in dimension by the radial basis function.
}
\details{
\code{rec} requires a vector to be converted to a array.
}
\examples{

imagedim1=c(10,10,10)
recvec = rec(rnorm(prod(imagedim1)), imagedim1)

}
