% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{data}
\name{template}
\alias{template}
\title{Brain Template}
\format{
A array
}
\usage{
data(template)
}
\description{
The data is the brain tempalte. This is an average brain image, and is mainly used for overlaying analysis results.
}
\keyword{datasets}
